Clean up MoveInService.cs
